requirements.txt
python-telegram-bot==20.6
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gspread==6.0.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
oauth2client==4.1.3
selenium==4.15.2
webdriver-manager==4.0.1
yfinance
requests
beautifulsoup4